requirements.txt
streamlit>=1.20.0
llama-index-llms-gemini>=0.1.4
llama-index-core>=0.10.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0